Create README.md
Hey everyone! This is my first dApp, that consists of a simple frontend, and I have deployed a Solidity smart contract, and connected them together.
This is an dApp that simply reads and writes a value to the blockchain. Here, we simply write a mood to the smart contract, and read the mood from the smart contract.